requirements.txt
openai
langchain
streamlit
python-dotenv
PyPDF2
pandas
langchain-community
langchain-openai